meta-amd: Enable OOB BMC firmware update
This patch enables phosphor-software-manager which updates BMC firmware.
Note that the version is based on Linux Kernel and is of the format:
"version=2.9.0-dev-52-gc09820583" saved in manifest file and only allows
updates if the versioning differs. While updating the developer builds
without using the Dediprog, MANIFEST file within *.image.tar would need
to be changed to a value different than the one running on the system.

TODO: To come up with AMD's versioning schema.

Change-Id: I68478bfc4425c902b91539fd1633d3b59fb440b3
Signed-off-by: Supreeth Venkatesh <supreeth.venkatesh@amd.com>