meta-amd: Modify device tree to enable thermal sensors for EthanolX
This patch modifies the device tree for EthanolX to correct the I2C bus
details on which the thermal sensors are present.

In Addition, this removes I2C1 to I2C8  bus details.
These I2C details have to be added as and when the corresponding
feature/functionality is enabled and tested.

Change-Id: I1b030bec2acd0a46ab4dac0d1291f827c46ca99a
Signed-off-by: Supreeth Venkatesh <supreeth.venkatesh@amd.com>