Update patch set 1
Patch Set 1:

Currently it is only possible to read a lower byte of the POST code. It is better than nothing, and that is what this commit is about.

To read full POST code we need some additional changes:
1) Enable second LPC snoop channel.
I've already pushed a patch for the Ethanol DTS to the Linux Kernel. 
I've added you as a reviewer, please check it out. Just in case, I'm talking about this:
https://lore.kernel.org/linux-arm-kernel/7d7a0990-c9ca-4113-aa41-415b5cb54cf5@www.fastmail.com/T/#t

2) 'phosphor-host-postd' needs some modifications:
I've pushed a change to `phosphor-host-postd` to read multibyte POST codes:
https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-host-postd/+/40064

3) If '2)' will get upstream, all is left to be done is add a simple override, for the `phosphor-host-postd` recipe:
```
SNOOP_DEVICE = "'aspeed-lpc-snoop0 aspeed-lpc-snoop1'"
```

I've tested all of the above on the EthanolX CRB, and it seems to work fine.

Patch-set: 1