Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Currently it is only possible to read a lower byte of the POST code. It is better than nothing, and that is what this commit is about.
> 
> To read full POST code we need some additional changes:
> 1) Enable second LPC snoop channel.
> I've already pushed a patch for the Ethanol DTS to the Linux Kernel. 
> I've added you as a reviewer, please check it out. Just in case, I'm talking about this:
> https://lore.kernel.org/linux-arm-kernel/7d7a0990-c9ca-4113-aa41-415b5cb54cf5@www.fastmail.com/T/#t
> 
> 2) 'phosphor-host-postd' needs some modifications:
> I've pushed a change to `phosphor-host-postd` to read multibyte POST codes:
> https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-host-postd/+/40064
> 
> 3) If '2)' will get upstream, all is left to be done is add a simple override, for the `phosphor-host-postd` recipe:
> ```
> SNOOP_DEVICE = "'aspeed-lpc-snoop0 aspeed-lpc-snoop1'"
> ```
> 
> I've tested all of the above on the EthanolX CRB, and it seems to work fine.

Post code is 32 bit and not 16 bit on EthanolX,so we will have to come up with a DMA driver to read all 32 bits. We have it working but it is not in a form to upstream it yet. I don't want to give out wrong reference by just adding 16 bit post code. I will wait for you if you have a complete solution. Display using 7 segment LED is 16 bit but the actual post code is 32 bit. i.e, port 80, 81, 82 and 83. Refer PPR document. Hence, it cant be enabled at this stage until the complete 32-bit post code is being read as it will be a violation of AMD PPR.

Patch-set: 1
Reviewer: Gerrit User 1000299 <1000299@adbd0a64-1f21-4d83-b585-671fe73cb6e4>